+++ NEWS.md
@@ -25,3 +25,4 @@
 * Documentation improvements.
 * `validate_country_nuts_countries` is now follows dplyr 1.0, this makes the code more readable.
 * `validate_nuts_regions` is validating non-EU NUTS-like regions as valid if they will be added to `NUTS2021`. These regional codes, while legally not part of the `NUTS2016` typology, are valid and can be placed on the maps created by [EuroGeographics maps provided by Eurostat](https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/nuts).
+* Estonia and Slovenia received pseudo-NUTS4 codes, and part of Latvia, too.